Card Structure Changes and Deck
Changed how cards are structured so that they each derive from a base scriptable object. Added a deck, which is given a list of each type of card and the amount of that card in the deck, and you can draw from the deck and shuffle it. Also basic debug functionality of printing out the cards in the deck with .ToString().